fix: 修复切换文档时I18N path不刷新 的问题
